Update phpcs
This is a major version bump which brings in several new sniffs and
updates many more. Some changes are accepted (e.g. improved spacing and
type union ordering, but many sniffs are ignored either due to
incomplete support of new PHP language features, or a preference for
more compact code.

* `PhpCollective.Commenting.Attributes.ExpectedFQCN` no longer exists.

* Accept standard reference assignment syntax instead of using the
  uncommon `=&` notation. As such, no longer suppress
  `PSR2R.WhiteSpace.UnaryOperatorSpacing.WrongSpace`.

+++ src/lib/Smr/Database.php
@@ -315,7 +315,7 @@ class Database {
 	 * @param T $num
 	 * @return T
 	 */
-	public function escapeNumber(int|float $num): int|float {
+	public function escapeNumber(float|int $num): float|int {
 		return $num;
 	}
